+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Added support for Terraria 1.4.4: Labour of Love.
+- Added support for Terraria 1.4.4: Labour of Love, superseding 1.4.3.
 - Added message notifying 1.4.3 users about Crowd Control for Terraria 1.4.4.
 - Added timed effect support in the Crowd Control browser source overlay.
 - Added "Use Moon-dial" effect; which fast forwards the time to dusk.